+++ plugins/codepipeline/frontend/src/components/CodePipelineExecutions/CodePipelineExecutions.tsx
@@ -11,7 +11,7 @@
  * limitations under the License.
  */
 
-import React, { useState } from 'react';
+import { useState } from 'react';
 import {
   PipelineExecutions,
   PipelineExecutionsResponse,